+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project are documented in this file.
 
 
 ## 2025-10-25
+### Fixed
+- Appbar Header now removes Header within TodoLists page
+
+
+## 2025-10-25
 ### Changes
 - Login now directs to first Workspace 
 ### Removed